requirements.txt
opencv-python==4.11.0.86
numpy==1.24.3
ultralytics==8.0.196
playsound==1.3.0
python-dotenv==1.0.0 